+++ candle-transformers/src/models/z_image/text_encoder.rs
@@ -399,17 +399,6 @@ impl ZImageTextEncoder {
         })
     }
 
-    /// Create causal attention mask
-    fn causal_mask(&self, b: usize, tgt: usize, offset: usize) -> Result<Tensor> {
-        let minf = f32::NEG_INFINITY;
-        let mask: Vec<_> = (0..tgt)
-            .flat_map(|i| {
-                (0..(tgt + offset)).map(move |j| if j <= i + offset { 0.0 } else { minf })
-            })
-            .collect();
-        Tensor::from_slice(&mask, (b, 1, tgt, tgt + offset), &self.device)?.to_dtype(self.dtype)
-    }
-
     /// Encode text, returning second-to-last layer hidden states
     ///
     /// # Arguments
@@ -420,13 +409,19 @@ impl ZImageTextEncoder {
     ///
     /// **Important**: Returns raw output from layer[-2] WITHOUT final RMSNorm
     pub fn forward(&self, input_ids: &Tensor) -> Result<Tensor> {
-        let (b, l) = input_ids.dims2()?;
+        let (_b, l) = input_ids.dims2()?;
         let mut hidden_states = self.embed_tokens.forward(input_ids)?;
 
         let causal = if l == 1 {
             None
         } else {
-            Some(self.causal_mask(b, l, 0)?)
+            Some(crate::utils::build_additive_causal_mask(
+                l,
+                0,
+                None,
+                &self.device,
+                self.dtype,
+            )?)
         };
 
         // num_hidden_layers = 36, second-to-last layer index = 34
